Agregando suma de control

+++ mainwindow.cpp
@@ -113,7 +113,7 @@ void MainWindow::onNewDataArrived(QStringList newData)
         int dataListSize = newData.size();
         uint16_t suma = 0, suma_recv=0;
         qDebug()<<"Lista = "<<newData;
-        suma_recv = newData[dataListSize-1];
+        suma_recv = newData[dataListSize-1].toLong;
         for(int i = 0; i < dataListSize-1; i++)
         {
             suma += newData[i].toLong();
